# log: Reduce allocated data in ContextKeyValues