In this change, we avoid copying the KVs from the Context until we know we need a new slice (i.e. we have something to add)

BenchmarkContextWith before and after

**before**
```
goos: darwin
goarch: arm64
pkg: github.com/luno/jettison/log
cpu: Apple M3 Pro
BenchmarkContextWith
BenchmarkContextWith-11    	 4461908	       254.6 ns/op	   3.93 MB/s	     464 B/op	       7 allocs/op
PASS
```
**after**
```
goos: darwin
goarch: arm64
pkg: github.com/luno/jettison/log
cpu: Apple M3 Pro
BenchmarkContextWith
BenchmarkContextWith-11    	 4701661	       224.6 ns/op	   4.45 MB/s	     400 B/op	       5 allocs/op
PASS
```