@andrewwormald 2025-03-07
Nice, improvement in allocs!